Revert "Topic information moved next to username"
